pyproject.toml
[project]
name = "audio-classifier"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.12",
    "huggingface-hub>=0.30.2",
    "librosa>=0.11.0",
    "matplotlib>=3.10.1",
    "pandas>=2.2.3",
    "python-dotenv>=1.1.0",
    "scikit-learn>=1.6.1",
    "torch>=2.6.0",
    "torchaudio>=2.6.0",
    "transformers>=4.51.3",
    "uvicorn>=0.34.2",
]
